Handle multiple input files
Running a test with multiple spec files would cause a new BrowserStack
tunnel to be started for each spec file. For projects with a large
number of spec files this causes an overflow error in Gulp.

It seems that the construction with the `tunnelStarted` flag was not
working correctly: the flag is always `false` when checked. The
startTunnel method now checks the `tunnel` object directly, solving the
issue and removing the need for a separate `tunnelStarted` flag.